+++ src/platform.ts
@@ -29,8 +29,6 @@ export class NoIPPlatform implements DynamicPlatformPlugin {
   platformConfig!: NoIPPlatformConfig
   platformLogging!: options['logging']
   platformRefreshRate!: options['refreshRate']
-  platformPushRate!: options['pushRate']
-  platformUpdateRate!: options['updateRate']
   platformAutoRenewal!: options['autoRenewal']
   platformRenewalInterval!: options['renewalInterval']
   version!: string
@@ -304,14 +302,6 @@ export class NoIPPlatform implements DynamicPlatformPlugin {
     this.platformRefreshRate = this.config.options?.refreshRate ? this.config.options.refreshRate : undefined
     const refreshRate = this.config.options?.refreshRate ? 'Using Platform Config refreshRate' : 'Platform Config refreshRate Not Set'
     await this.debugLog(`${refreshRate}: ${this.platformRefreshRate}`)
-    // UpdateRate
-    this.platformUpdateRate = this.config.options?.updateRate ? this.config.options.updateRate : undefined
-    const updateRate = this.config.options?.updateRate ? 'Using Platform Config updateRate' : 'Platform Config updateRate Not Set'
-    await this.debugLog(`${updateRate}: ${this.platformUpdateRate}`)
-    // PushRate
-    this.platformPushRate = this.config.options?.pushRate ? this.config.options.pushRate : undefined
-    const pushRate = this.config.options?.pushRate ? 'Using Platform Config pushRate' : 'Platform Config pushRate Not Set'
-    await this.debugLog(`${pushRate}: ${this.platformPushRate}`)
   }
 
   async getPlatformRenewalSettings() {
@@ -332,8 +322,6 @@ export class NoIPPlatform implements DynamicPlatformPlugin {
       }
       platformConfig.logging = this.config.options.logging ? this.config.options.logging : undefined
       platformConfig.refreshRate = this.config.options.refreshRate ? this.config.options.refreshRate : undefined
-      platformConfig.updateRate = this.config.options.updateRate ? this.config.options.updateRate : undefined
-      platformConfig.pushRate = this.config.options.pushRate ? this.config.options.pushRate : undefined
       platformConfig.autoRenewal = this.config.options.autoRenewal ? this.config.options.autoRenewal : undefined
       platformConfig.renewalInterval = this.config.options.renewalInterval ? this.config.options.renewalInterval : undefined
       if (Object.entries(platformConfig).length !== 0) {